feat: Added ledgerconfig update sub-command
Implemented the update sub-command to add/update the configuration of one or more applications.
Configuration can be specified directly on the command-line as a JSON string using the --config option,
or the path of a configuration file may be specified using the --configfile option. The configuration
string may be embedded directly in the "Config" element or the Config element may reference a file
containing the configuration.

closes #3

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>