feat: Added ledgerconfig delete sub-command
Implemented the delete sub-command to delete the MSP's configuration using search criteria. The criteria consists of:

- MspID: The MSP ID of the organization
- PeerID: The ID of the peer
- AppName: The application name
- AppVersion: The application version
- ComponentName: The component name
- ComponentVersion: The component version

Criteria may be specified as a JSON string (using the --criteria option) or it may be specified using the options:
	--mspid, --peerid, --appname, --appver, --componentname and --componentver

If PeerID and AppName are not specified then all of the MSP's configuration is deleted.

closes #4

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>